fastapi/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.32
alembic>=1.13.2
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
email-validator>=2.2.0
psycopg[binary]>=3.2.0
